Fix package scope command
The migration to `@stylistic/stylelint-config` surfaced a flaw in the
GitHub Packages publish job. The `sed` command was too liberal and
replaced `@stylistic` with `@jgarber623` in the package file's
dependencies. That caused the `npm ci` command to fail as shown in the
logs below.

This change addresses that by updating the `sed` command.

See logs from failed publish run:

https://github.com/jgarber623/stylelint-config/actions/runs/7439726730/job/20239936874